+++ app/Models/Filter/EventFilter.php
@@ -2,14 +2,15 @@
 
 namespace App\Models\Filter;
 
+use App\Models\MemberGroup;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Request;
 
 class EventFilter
 {
-    public const PARAM_INCL_DISABLED = 'disabled';
+    public const PARAM_MEMBER_GROUP = 'memberGroup';
 
-    public const PARAM_INCL_LOGGED_IN_ONLY = 'loggedInOnly';
+    public const PARAM_INCL_DISABLED = 'disabled';
 
     public const PARAM_START_DATE = 'start';
 
@@ -19,21 +20,27 @@ class EventFilter
 
     public bool $inclDisabled;
 
-    public bool $inclLoggedInOnly;
+    /**
+     * @var MemberGroup[]
+     */
+    public array $memberGroups;
 
     public ?Carbon $start;
 
     public ?Carbon $end;
 
+    /**
+     * @param  MemberGroup[]  $memberGroups
+     */
     public function __construct(
         Carbon $start = null,
         Carbon $end = null,
         bool $inclDisabled = false,
-        bool $inclLoggedInOnly = false,
+        array $memberGroups = [],
         bool $sortAsc = true)
     {
         $this->inclDisabled = $inclDisabled;
-        $this->inclLoggedInOnly = $inclLoggedInOnly;
+        $this->memberGroups = $memberGroups;
         $this->sortAsc = $sortAsc;
         $this->start = $start;
         $this->end = $end;
@@ -45,7 +52,10 @@ class EventFilter
             self::PARAM_START_DATE => $this->start,
             self::PARAM_END_DATE => $this->end,
             self::PARAM_INCL_DISABLED => $this->inclDisabled,
-            self::PARAM_INCL_LOGGED_IN_ONLY => $this->inclLoggedInOnly,
+            self::PARAM_MEMBER_GROUP => implode(',', array_map(
+                fn (MemberGroup $mg) => $mg->id,
+                $this->memberGroups
+            )),
         ];
     }
 
@@ -55,6 +65,8 @@ class EventFilter
             Request::get(self::PARAM_START_DATE),
             Request::get(self::PARAM_END_DATE),
             Request::get(self::PARAM_INCL_DISABLED),
-            Request::get(self::PARAM_INCL_LOGGED_IN_ONLY));
+            MemberGroup::query()->find(
+                explode(',', Request::get(self::PARAM_MEMBER_GROUP))
+            )->toArray());
     }
 }