+++ resources/views/components/livewire/filter/event-filter.blade.php
@@ -1,10 +1,5 @@
 <div x-data="{
-    open:false,
-    start: $persist(@entangle('start').live).using(cookieStorage),
-    end: $persist(@entangle('end').live).using(cookieStorage),
-    showDisabled: $persist(@entangle('showDisabled').live),
-    showLoggedInOnly: $persist(@entangle('showLoggedInOnly').live),
-    sortAsc: $persist(@entangle('sortAsc').live),
+    open:false
 }" class="relative text-left" @click.outside="open = false">
 
 
@@ -29,12 +24,18 @@
                 </x-input-checkbox>
             </div>
         @endif
-        @if($this->canFilterShowLoggedInOnly())
+        @if($this->canFilterMemberGroup())
             <div class="px-4 py-1">
-                <x-input-checkbox id="filter_logged_in_only" name="filter_logged_in_only"
-                                  wire:model.live="showLoggedInOnly">
-                    {{ __('show logged in only events') }}
-                </x-input-checkbox>
+                <x-input-label for="member_group_id" :value="__('Member group')"/>
+                <x-select id="member_group_id" name="member_group_id"
+                          wire:model.live="memberGroupId"
+                          class="block mt-1 w-full">
+                    <option value="all">All</option>
+                    <option value="">-</option>
+                    @foreach(\App\Models\MemberGroup::getTopLevelQuery()->get() as $memberGroup)
+                        <x-members.member-group-select-option :memberGroup="$memberGroup"/>
+                    @endforeach
+                </x-select>
             </div>
         @endif
         <x-accordion label="Date filter" class="min-w-60 text-sm text-gray-700">
@@ -47,8 +48,7 @@
                 @if(!$this->isStartNow)
                     <label title="{{__('Filter start date')}}"
                            class="flex border border-gray-700 divide-x focus:border-indigo-500 focus:ring-indigo-500 rounded-md shadow-sm m-1 w-fit overflow-hidden">
-                        <div class="px-2 flex items-center"
-                             :class="{ 'bg-green-700': start, 'bg-gray-500': !start }">
+                        <div class="px-2 flex items-center {{ $this->start ? 'bg-green-700' : 'bg-gray-500' }}">
                             <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24">
                                 <path fill="currentColor"
                                       d="M2 14h2v3h5v-3l4 4l-4 4v-3H4v3H2zm17 5V8H5v4H3V5c0-1.11.89-2 2-2h1V.998h2V3h8V.998h2V3h1a2 2 0 0 1 2 2v14c0 1.1-.9 2-2 2h-6.17l2-2z"/>
@@ -61,8 +61,7 @@
             <div class="">
                 <label title="{{__('Filter end date')}}"
                        class="flex border border-gray-700 divide-x focus:border-indigo-500 focus:ring-indigo-500 rounded-md shadow-sm m-1 w-fit overflow-hidden">
-                    <div class="px-2 flex items-center"
-                         :class="{ 'bg-red-700': end, 'bg-gray-500': !end }">
+                    <div class="px-2 flex items-center {{ $this->end ? 'bg-red-700' : 'bg-gray-500' }}">
                         <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24">
                             <path fill="currentColor"
                                   d="M22 14v8h-2v-4l-4 4v-3h-5v-2h5v-3l4 4v-4zM5 19h4v2H5c-1.1 0-2-.9-2-2V5a2 2 0 0 1 2-2h1V.998h2V3h8V.998h2V3h1c1.11 0 2 .89 2 2v7h-2V8H5z"/>